Add compatibility with print()-style syntax
The new logroll.printer(depth) returns a function which then accepts the same
syntax as lua's (and torch's) print() function, and can thus be used as
a drop-in replacement.

Use-case: in combination with rosejn/logroll, to output nicely to a
file in a simlar fashion to printing on the screen.

Joint work with @koraykv (Thanks!)